Refactoring: moving services related code into separate packages
