+++ benchmarks/src/main/java/sustain/synopsis/benchmarks/dht/writethroughput/local/Driver.java
@@ -4,7 +4,8 @@ import com.google.protobuf.ByteString;
 import io.grpc.BindableService;
 import org.apache.log4j.Logger;
 import sustain.synopsis.dht.*;
-import sustain.synopsis.dht.services.IngestionService;
+import sustain.synopsis.dht.services.ingestion.IngestionRequestDispatcher;
+import sustain.synopsis.dht.services.ingestion.IngestionService;
 import sustain.synopsis.dht.store.StorageException;
 import sustain.synopsis.dht.store.node.NodeStore;
 import sustain.synopsis.dht.store.services.IngestionRequest;
