+++ dht/src/main/java/sustain/synopsis/dht/services/query/Interval.java
@@ -1,4 +1,4 @@
-package sustain.synopsis.dht.store.query;
+package sustain.synopsis.dht.services.query;
 
 import java.util.Objects;
