+++ dht/src/main/java/sustain/synopsis/dht/services/query/MatchedSSTable.java
@@ -1,4 +1,4 @@
-package sustain.synopsis.dht.store.query;
+package sustain.synopsis.dht.services.query;
 
 import sustain.synopsis.dht.store.StrandStorageKey;
 import sustain.synopsis.storage.lsmtree.Metadata;
