+++ dht/src/test/java/sustain/synopsis/dht/store/entity/EntityStoreTest.java
@@ -7,9 +7,9 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import sustain.synopsis.dht.store.*;
-import sustain.synopsis.dht.store.query.Interval;
-import sustain.synopsis.dht.store.query.MatchedSSTable;
-import sustain.synopsis.dht.store.query.QueryException;
+import sustain.synopsis.dht.services.query.Interval;
+import sustain.synopsis.dht.services.query.MatchedSSTable;
+import sustain.synopsis.dht.services.query.QueryException;
 import sustain.synopsis.dht.store.services.Expression;
 import sustain.synopsis.dht.store.services.Predicate;
 import sustain.synopsis.storage.lsmtree.Metadata;
